feat(engine): support running wasm-based plugins
This commit introduces the building blocks to run game play logic
through Wasm modules.

The commit introduces two relevant concepts:

1. A `Plugin` encapsulates a Wasm module loaded into memory.

2. The `PluginManager` is responsible for loading and running `Plugin`s.

Part of this commit is an effort to have the engine return structured
and relevant error messages.

The `Engine::run` method currently loads a single test plugin to
validate the functionality, but this will be migrated to a more flexible
set-up to allow changing which plugins get loaded by the `runner`.