feat(plugin-test): add Rust-based Wasm plugin
Adding a Rust-based Wasm plugin will help evolve the API used to
communicate between the engine and future plugins.

The goal is to deliver an SDK to make it easy for plugins to be written
in Rust and compiled to Wasm.

This might clear the path for SDKs in other languages in the future.

For now, there's a single `test` plugin, but the goal is to have as much
of the core game play logic be written as a set of plugins that ship
with the default package, allowing others to extend the functionality of
a game with custom plugins.

The engine would be responsible for running plugins, and feeding them
the information provided by the player required to drive the story
forward.

But for now... The plugin is called "test" for a reason. Let's start
small.