feat(runner): run all rust-based wasm plugins
The runner now iterates over all plugins in the `plugins` directory.
This means it will pick up any Rust plugins that are compiled to wasm.

The process of testing a plugin change is now:

    cargo build --lib --target wasm32-unknown-unknown --manifest-path plugins/test/Cargo.toml
    cargo run

The path search logic removes wasm modules with duplicate names. While
this likely shouldn't remain a constraint, for now this makes things
easier, as Rust will compile the plugins into two places in the `target`
directory, which would result in the plugins being ran twice.