feat(sdk): introduce SDK crate
This crate serves as the main entrypoint to developing plugins in Rust.

The purpose of this crate is to abstract away any intrinsic details
about working with foreign function interfaces and pointer logic to pass
data between the engine and the plugin.

The eventual goal is to provide one macro `vienna::load!()` which, when
added, will guide you towards implementing the required plugin functions
using regular types, as the Rust compiler warns you about missing or
invalid function implementations.

For now, it implements the existing `_run` FFI function, and requires
the implementor to add a `run` function to satisfy the SDK requirements.