refactor(engine): move plugin instantiation into plugin
The plugin handler now takes a path to a plugin, reads its bytes, and
then hands off plugin initialization to the `Plugin`'s `new` method.

This creates a more appropriate seam between what the handler is
responsible for, and the plugin.

This also prepares the plugin to be able to run the exported `init`
function to get plugin details before initialisation.