feat(engine): add plugin initialization support
In order for plugins to work as expected, they need some way to
communicate to the engine how they work, and what they expect from the
engine.

For example, a plugin should be able to advertise its name, or the
events it is interested in. When more security-focused features are
added, plugins should also be able to ask the engine for specific access
rights, which the engine can then either accept or reject.

Traditionally, game mods expose this information through a configuration
file.

In Vienna, the plugin exposes this information at runtime when the
plugin is initialized.

This has several advantages, amongst other a plugin continues to be a
single file, without the need to pack/unpack multiple files into one. It
also makes developing plugins more convenient as all configuration lives
in code, and can be set in a more dynamic way.

---

As a start, each plugin exposes its name to the engine.

As for the implementation, the call stack looks something like this:

- plugin path is provided to `Handler::register_plugin`
- handler reads the file and provides the file contents to `Runtime::new`
  - the bytes are converted to a WebAssembly instance
  - the `_init` Wasm function is called
    - the Wasm plugin creates a `Registration` struct
    - the struct is serialized to JSON (for now)
    - the plugin calls back to the engine with the memory pointer and
      string length
  - the engine reads the plugin's memory to re-construct the
    `Registration` struct
  - this struct is used to configure the plugin
  - `Runtime::new` either returns an error if the initialization failed,
    or the plugin is constructed as expected
- the handler stores the plugin in its memory
- `Runtime#run` is now available to the handler to run the plugin

The serialization to/from JSON provides some overhead to the current
implementation. The next steps will be to experiment with `#[repr(C)]`
to pass raw pointers from the plugin to the engine to avoid the extra
serialization/deserialization step.