feat(engine): add initial game loop and window
This PR integrates the `ggez` crate to provide a functional game loop,
rendering engine and windowing system.

This is still an early experimental implementation, and the intention is
to keep ggez types as self-contained as possible.

The reason for this is that it's not clear yet if using ggez is the
right choice, or perhaps something like quicksilver[0] would be a better
choice.

For now, the focus is on getting a fully integrated loop going,
consisting of:

- window management
- player input
- plugin execution
- state mutation
- state rendering

The next step will be to have a plugin manipulate the demo `pos_x`
state. After that, we'll convert the current circle-shape into a
"widget" (see RFC#3) and have the plugin manage its life-cycle.

[0]: https://github.com/ryanisaacg/quicksilver