feat(engine): send events to plugins
This is an initial proof of concept for a system that generates `Event`s
and sends them to plugins to handle.

The current implementation supports keyboard events for the WASD keys.
As a test, the `test` plugins takes these input events and updates the
`pos_x` and `pos_y` variables, which the renderer uses to render a
circle on the screen.

This allows the player to move the circle on the screen using the WASD
keys.

While this works, the eventual goal is not to have plugins deal with raw
input values, but instead translate them to specific "widget"
operations (see RFC#004).