refactor(engine): move ggez impl into separate module
This commit helps isolate the "ggez" game engine types into a dedicated
module.

The goal is to have as little dependencies as possible on any specific
game engine implementation.

Currently, the `renderer` module still uses render-specific types of
ggez, but this is a temporary situation until plugins are in charge of
driving the screen.