feat(engine): integrate "coffee" engine
Vienna now supports two separate game engines.

These are "ggez"[0] (the one already present before this commit), and
"coffee"[1], a new, more lightweight engine.

These two engines are together named "core", and live in the engine
crate, in the `core` module. Since using both at the same time is
pointless, they are each put behind a feature flag, `core-ggez` and
`core-coffee`, one of both has to be enabled.

By default, `core-coffee` is enabled.

The goal of this exercise is to see if we can implement an
engine-agnostic framework within Vienna. By doing this, we should (in
theory) be able to add more unit testing without relying on a windowing
framework, by adding a (future) third core; one that uses a terminal,
and can therefor run "headless" during testing.

The cores are not entirely encapsulated yet, there are still a few
sprinkles of `cfg`s throughout the code-base to change behaviour based
on the enabled core.

Also, for now both cores have their own updaters and renderers, although
the goal is to move as much as possible into a shared updater/renderer,
and then call `core::updater` and `core::renderer` to trigger
core-specific logic.

[0]: https://ggez.rs
[1]: https://docs.rs/coffee