feat(engine): introduce "generic widgets"
This builds on top of c2b12b5 and implements recent changes to in RFC004
introduced in d232886 and the new RFC005 added in cc0b21a.

Specifically, this commit converts to a generic approach for widgets,
which allows new widgets to be added in the engine, and opens the door
for WebAssembly-based widgets. The latter is not implemented yet, but
these changes are needed for that to happen.

In short, the following flow of steps and events happen:

1. A widget implements the `widget::Runtime` trait.
2. Plugin defines one or more widget instances it wants to manage.
3. Game is started.
4. Engine registers plugin, instantiating the relevant widgets.
5. Player triggers "input event" (keyboard, mouse, etc.).
6. Engine updates the game state.
   - All registered widgets have their `interact` method called.
     - Widgets update their state based on the player interaction.
     - Widgets emit one or more "widget events".
   - Plugins run.
     - Plugin gets a list of emitted widget events.
     - Plugin can access and manipulate the widgets it owns.
7. Engine renders the game state.
   - All registered widgets have their `draw` method called.
     - Widget return a list of "primitive shapes" based on its internal
       state.
   - Engine draws the list of primitive shapes.
8. Go to step 5.

The biggest implication for plugins is that they no longer act on raw
"input events", but instead act on widget events. Eventually, other
types of events will be added which the plugin can receive (such as
timer events), but player input is hidden from plugins and instead
interpreted by widgets.

This allows for a clear separation between games, widgets, and plugins,
making it possible to create generic plugins that work for multiple
games, regardless of how they work.