feat(engine): support WebAssembly System Interface plugins
This allows plugins to tap into OS resources such as random entropy to
support randomisation in plugins.

It also allows for easier debugging by printing to stdout from plugins.

In the future, this will either be removed, or restricted to a subset of
capabilities, to prevent mis-use of game plugins.