feat(widget): add `is_within_bounds` method
The engine will use this method to ask the plugin if a given coordinate
falls within the "bounds" of the widget.

By default, this method always returns `true`, because the engine
guarantees to only call this method if the coordinate falls between the
top-left and bottom-right coordinate of the boxed widget, as defined by
the existing `dimensions()` method each widget implements.

By overriding this method, a widget can redefine the bounds of the
widget based on its own shapes.

For example, a widget in the shape of a circle, can use its radius and
the Pythagorean theorem to give exact details on the edge of the widget.

Bounds checking is currently used to determine if a specific input event
should be sent to the widget (e.g. "hover", "blur" and "pointer
movement"), but can be extended in the future for collision detection.