feat(engine): add initial mouse support
Widgets can now receive mouse/cursor events.

To start with, there are four new events:

- `Pointer` (coordinates of the location of the cursor)
- `MouseClick` (coordinates and the button that was clicked)
- `Focus` (the mouse enters the bounds of a widget)
- `Blur` (the mouse leaves the bounds of a widget)

All of these events are related to a specific _widget_. Meaning, the
coordinates attached to these events are relative to the widget
(starting at 0,0 for the top-left bound of the widget).

This uses the new `is_within_bounds()` method introduced in 07b3164 to
check if a mouse event is related to a given widget.