feat(engine): initial support for "drag" events
This is a very crude first implementation to allow widgets to detect
long-press events, which can then be translated into "drag" actions.

As a proof of concept, the "moving circle" widget allows for pressing
down the left mouse button on it, after which the circle will follow the
mouse pointer as if it were being dragged around.

There are still many flaws in the implementation:

- The drag motion not as smooth as it could be.
- If you don't drag from the center, the first drag event will make the
  circle jump to center on the cursor.
- The drag detection isn't great.
- If the cursor moves out of the circle due to fast movement, dragging
  stops.

Most importantly though, this exposes some issues in the current event
handling logic that requires a larger refactor to make it more correct,
and ergonomic for widget and plugin authors.

There will be an RFC to discuss the needed changes to the event system.

For now though, it "works", and is a good way to stress-test the current
systems.